% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARPALData.R
\docType{package}
\name{ARPALData}
\alias{ARPALData-package}
\alias{ARPALData}
\title{ARPALData Package}
\description{
Contains functions for downloading and managing air quality and weather data from Regione Lombardia open database.
    Data are collected by ARPA Lombardia (Lombardia Environmental Protection Agency), Italy.
}
\author{
Paolo Maranzano \email{pmaranzano.ricercastatistica@gmail.com}
}
