% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ARPA_Lombardia_AQ_municipal_registry.R
\name{get_ARPA_Lombardia_AQ_municipal_registry}
\alias{get_ARPA_Lombardia_AQ_municipal_registry}
\title{Download metadata (registry) on air quality monitoring stations at municipal level from ARPA Lombardia website}
\usage{
get_ARPA_Lombardia_AQ_municipal_registry()
}
\value{
A data frame of class 'data.frame' and 'ARPALdf'. The object is fully compatible with Tidyverse.
}
\description{
'get_ARPA_Lombardia_AQ_municipal_registry' returns the registry (list) of all the air quality sensors
owned by ARPA Lombardia for each municipality of Lombardy. The information reported are: ID of each sensor and
station, starting date and ending date. The column 'NameStation' identifies the name of each municipality.
The column 'IDStation' is an ID code (assigned from ARPA) uniquely identifying each municipality.
For more information about the municipal data visit the section 'Stime comunali sull'aria' at the webpage:
https://www.dati.lombardia.it/stories/s/auv9-c2sj
}
\examples{
get_ARPA_Lombardia_AQ_municipal_registry()

}
