% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ARPA_Lombardia_AQ_registry.R
\name{get_ARPA_Lombardia_AQ_registry}
\alias{get_ARPA_Lombardia_AQ_registry}
\title{Download metadata (registry) on air quality monitoring stations from ARPA Lombardia website}
\usage{
get_ARPA_Lombardia_AQ_registry()
}
\value{
A data frame of class 'data.frame' and 'ARPALdf'. The object is fully compatible with Tidyverse.
}
\description{
'get_ARPA_Lombardia_AQ_registry' returns the registry (list) of all the air quality sensors
and stations belonging to the ARPA Lombardia network. The information reported are: ID of each sensor and
station, geo-location (coordinates in degrees), altitude (mt), starting date and ending date.
The column 'NameStation' identifies the name of each station, while 'IDStation' is an ID code (assigned
from ARPA) uniquely identifying each station.
For more information about the municipal data visit the section 'Monitoraggio aria' at the webpage:
https://www.dati.lombardia.it/stories/s/auv9-c2sj
}
\examples{
get_ARPA_Lombardia_AQ_registry()

}
