% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ARPALdf.R
\name{is_ARPALdf}
\alias{is_ARPALdf}
\title{Check if a given object belongs to class 'ARPALdf'}
\usage{
is_ARPALdf(Data)
}
\arguments{
\item{Data}{Object to check if the class of a dataframe is 'ARPALdf', i.e. ARPAL dataframe.}
}
\value{
The function returns 'True' if the object is of class 'ARPALdf' and it returns 'False' if the
    object isn't of class 'ARPALdf'
}
\description{
'is_ARPALdf' checks if the input object belongs to the class 'ARPALdf'
}
\examples{
d <- get_ARPA_Lombardia_AQ_registry()
is_ARPALdf(d)

}
