% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ARPALdf_W.R
\name{is_ARPALdf_W}
\alias{is_ARPALdf_W}
\title{Check if a given object belongs to class 'ARPALdf_W'}
\usage{
is_ARPALdf_W(Data)
}
\arguments{
\item{Data}{Object to check if the class of a dataframe is 'ARPALdf_W', i.e. ARPAL
dataframe for weather data.}
}
\value{
The function returns 'True' if the object is of class 'ARPALdf_W' and it returns 'False' if the
    object isn't of class 'ARPALdf_W'
}
\description{
'is_ARPALdf_W' checks if the input object belongs to the class 'ARPALdf_W'
}
\examples{
d <- get_ARPA_Lombardia_W_registry()
is_ARPALdf_W(d)

}
